add license and notice to kwik extractor
